agoda/Solution/executable/requirement.txt
xlrd==1.2.0
numpy==1.19.1 
pandas==1.1.0
python-Levenshtein==0.12.0
fuzzywuzzy==0.18.0
